Split agent answer formatter families
