+++ Sources/QuillCodeAgent/AgentFinalAnswerBuilder.swift
@@ -36,10 +36,3 @@ enum AgentFinalAnswerBuilder {
         return "Output:\n\(AgentToolAnswerFormatters.truncated(output))"
     }
 }
-
-private extension String {
-    var trimmedNonEmpty: String? {
-        let trimmed = trimmingCharacters(in: .whitespacesAndNewlines)
-        return trimmed.isEmpty ? nil : trimmed
-    }
-}
