+++ Tests/QuillCodeParityTests/ParityAgentGateTests.swift
@@ -5,6 +5,11 @@ final class ParityAgentGateTests: QuillCodeParityTestCase {
         let agentText = try Self.agentSourceText(named: "Agent.swift")
         let builderText = try Self.agentSourceText(named: "AgentFinalAnswerBuilder.swift")
         let formatterText = try Self.agentSourceText(named: "AgentToolAnswerFormatters.swift")
+        let supportText = try Self.agentSourceText(named: "AgentToolAnswerFormatterSupport.swift")
+        let shellFormatterText = try Self.agentSourceText(named: "AgentShellToolAnswerFormatters.swift")
+        let browserFormatterText = try Self.agentSourceText(named: "AgentBrowserToolAnswerFormatters.swift")
+        let gitFormatterText = try Self.agentSourceText(named: "AgentGitToolAnswerFormatters.swift")
+        let utilityFormatterText = try Self.agentSourceText(named: "AgentUtilityToolAnswerFormatters.swift")
 
         XCTAssertTrue(builderText.contains("enum AgentFinalAnswerBuilder"), "Tool-result final answer copy should live in a focused builder.")
         XCTAssertTrue(builderText.contains("static func finalAnswer"), "Final answer formatting should be directly testable.")
@@ -12,8 +17,17 @@ final class ParityAgentGateTests: QuillCodeParityTestCase {
         XCTAssertFalse(builderText.contains("import QuillCodeTools"), "The builder should not import tool modules once formatting has a separate owner.")
         XCTAssertTrue(formatterText.contains("enum AgentToolAnswerFormatters"), "Tool-specific final-answer summaries should live in a dedicated formatter owner.")
         XCTAssertTrue(formatterText.contains("static var all: [Formatter]"), "Tool-specific final answers should be registered instead of growing the top-level finalAnswer method.")
-        XCTAssertTrue(formatterText.contains("ToolDefinition.shellRun.name"), "Shell final-answer special cases should live with tool formatters.")
-        XCTAssertTrue(formatterText.contains("ToolDefinition.browserInspect.name"), "Browser final-answer special cases should live with tool formatters.")
+        XCTAssertTrue(formatterText.contains("AgentShellToolAnswerFormatters.shellRunAnswer"), "The registry should delegate shell copy to a focused formatter.")
+        XCTAssertTrue(formatterText.contains("AgentBrowserToolAnswerFormatters.browserInspectAnswer"), "The registry should delegate browser copy to a focused formatter.")
+        XCTAssertTrue(formatterText.contains("AgentGitToolAnswerFormatters.pullRequestReviewThreadsAnswer"), "The registry should delegate PR review-thread copy to a focused formatter.")
+        XCTAssertTrue(formatterText.contains("AgentUtilityToolAnswerFormatters.memoryRememberAnswer"), "The registry should delegate utility tool copy to a focused formatter.")
+        XCTAssertTrue(supportText.contains("enum AgentToolAnswerFormatterSupport"), "Shared formatter argument parsing should stay out of domain formatter files.")
+        XCTAssertTrue(shellFormatterText.contains("ToolDefinition.shellRun.name"), "Shell final-answer special cases should live with shell formatters.")
+        XCTAssertTrue(browserFormatterText.contains("ToolDefinition.browserInspect.name"), "Browser final-answer special cases should live with browser formatters.")
+        XCTAssertTrue(gitFormatterText.contains("ToolDefinition.gitPullRequestReviewThreads.name"), "PR review-thread final-answer copy should live with git formatters.")
+        XCTAssertTrue(utilityFormatterText.contains("ToolDefinition.memoryRemember.name"), "Memory final-answer copy should live with utility formatters.")
+        XCTAssertFalse(formatterText.contains("private static func shellAnswer"), "The registry should not own shell-specific formatting logic.")
+        XCTAssertFalse(formatterText.contains("PullRequestReviewThreadsResponse"), "The registry should not own PR response models.")
         XCTAssertTrue(agentText.contains("AgentFinalAnswerBuilder.finalAnswer"), "AgentRunner should delegate final-answer formatting.")
         XCTAssertFalse(agentText.contains("private static func shellAnswer"), "AgentRunner should not own shell final-answer formatting.")
         XCTAssertFalse(agentText.contains("private static func browserInspectionAnswer"), "AgentRunner should not own browser final-answer formatting.")
